Add Rodi supplemental algorithm suitable for nodal-based assembly
* Add Rodi source term

* Add unit test for Rodi; augmented the Turbulence initializations.

* Add documentation for Rodi in theory manual

Usage:
        - source_terms:
            turbulent_ke: rodi

Notes:

a) Only a unit test for coverage. Creation verified in a side case.

+++ unit_tests/UnitTestUtils.C
@@ -183,7 +183,7 @@ stk::mesh::Entity create_one_element(
    stk::mesh::EntityIdVector nodeIds(topo.num_nodes());
    std::iota(nodeIds.begin(), nodeIds.end(), 1);
 
-
+   
    bulk.modification_begin();
 
    for (auto id : nodeIds) {
@@ -469,7 +469,7 @@ double vector_norm(const std::vector<double> & vec, const stk::ParallelMachine& 
   double norm = 0.0;
   double g_norm = 0.0;
 
-  for (int i = 0; i < N; ++i) {
+  for (size_t i = 0; i < N; ++i) {
     norm += vec[i]*vec[i];
   }
   stk::all_reduce_sum(comm, &N, &g_N, 1);
